@@ -1,11 +1,9 @@
 /*!
- * Copyright (c) 2021-2022 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import '@bedrock/express';
 
-// translate `main.js` to CommonJS
-require = require('esm')(module);
-module.exports = require('./main.js');
+import './http.js';
 
-// backwards compatibility: export full API as `default`
-module.exports.default = module.exports;
+// load config defaults
+import './config.js';
