@@ -1,10 +1,12 @@
 /*
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import * as bedrock from 'bedrock';
+import * as bedrock from '@bedrock/core';
 import * as helpers from './helpers.js';
-import {KeystoreConfigStorage} from '@digitalbazaar/webkms-switch';
-import {keystores} from 'bedrock-kms';
+import {createRequire} from 'module';
+import {keystores} from '@bedrock/kms';
+const require = createRequire(import.meta.url);
+const {KeystoreConfigStorage} = require('@digitalbazaar/webkms-switch');
 
 const {util: {BedrockError}} = bedrock;
 
