@@ -1,10 +1,10 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import {config} from 'bedrock';
-import 'bedrock-validation';
-import 'bedrock-app-identity';
-import 'bedrock-meter-usage-reporter';
+import {config} from '@bedrock/core';
+import '@bedrock/app-identity';
+import '@bedrock/meter-usage-reporter';
+import '@bedrock/validation';
 
 const namespace = 'kms-http';
 const cfg = config[namespace] = {};
