@@ -1,18 +1,18 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import * as bedrock from 'bedrock';
-import * as brZCapStorage from 'bedrock-zcap-storage';
+import * as bedrock from '@bedrock/core';
+import * as brZCapStorage from '@bedrock/zcap-storage';
 import * as helpers from './helpers.js';
 import * as middleware from './middleware.js';
-import {asyncHandler} from 'bedrock-express';
+import {asyncHandler} from '@bedrock/express';
 import cors from 'cors';
-import {createValidateMiddleware as validate} from 'bedrock-validation';
+import {createRequire} from 'module';
+import {createValidateMiddleware as validate} from '@bedrock/validation';
 import {
   defaultModuleManager as moduleManager,
   keystores
-} from 'bedrock-kms';
-import {generateRandom} from '@digitalbazaar/webkms-switch';
+} from '@bedrock/kms';
 import {meters} from 'bedrock-meter-usage-reporter';
 import {
   postKeystoreBody,
@@ -20,6 +20,8 @@ import {
   postRevocationBody
 } from '../schemas/bedrock-kms-http.js';
 import {reportOperationUsage, SERVICE_TYPE} from './metering.js';
+const require = createRequire(import.meta.url);
+const {generateRandom} = require('@digitalbazaar/webkms-switch');
 
 const {config, util: {BedrockError}} = bedrock;
 
