@@ -1,5 +1,6 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import {loggers} from 'bedrock';
+import {loggers} from '@bedrock/core';
+
 export const logger = loggers.get('app').child('bedrock-kms-http');
