@@ -1,13 +1,13 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import * as bedrock from 'bedrock';
-import * as brZCapStorage from 'bedrock-zcap-storage';
+import * as bedrock from '@bedrock/core';
+import * as brZCapStorage from '@bedrock/zcap-storage';
 import {
   defaultModuleManager as moduleManager,
   keystores
-} from 'bedrock-kms';
-import {meters} from 'bedrock-meter-usage-reporter';
+} from '@bedrock/kms';
+import {meters} from '@bedrock/meter-usage-reporter';
 import {logger} from './logger.js';
 
 // configure usage aggregator for webkms meters
