@@ -1,25 +1,28 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import * as bedrock from 'bedrock';
-import * as brZCapStorage from 'bedrock-zcap-storage';
+import * as bedrock from '@bedrock/core';
+import * as brZCapStorage from '@bedrock/zcap-storage';
 import * as helpers from './helpers.js';
-import {asyncHandler} from 'bedrock-express';
-import {
-  authorizeZcapInvocation as _authorizeZcapInvocation,
-  authorizeZcapRevocation as _authorizeZcapRevocation
-} from '@digitalbazaar/ezcap-express';
+import {asyncHandler} from '@bedrock/express';
 import {BedrockKeystoreConfigStorage} from './BedrockKeystoreConfigStorage.js';
-import {createOperationMiddleware} from '@digitalbazaar/webkms-switch';
-import {CryptoLD} from 'crypto-ld';
+import {createRequire} from 'module';
 import {
   defaultDocumentLoader as documentLoader,
   defaultModuleManager as moduleManager
-} from 'bedrock-kms';
-import {Ed25519Signature2020} from '@digitalbazaar/ed25519-signature-2020';
-import {Ed25519VerificationKey2020} from
-  '@digitalbazaar/ed25519-verification-key-2020';
+} from '@bedrock/kms';
 import {reportOperationUsage} from './metering.js';
+const require = createRequire(import.meta.url);
+const {
+  authorizeZcapInvocation: _authorizeZcapInvocation,
+  authorizeZcapRevocation: _authorizeZcapRevocation
+} = require('@digitalbazaar/ezcap-express');
+const {createOperationMiddleware} = require('@digitalbazaar/webkms-switch');
+const {CryptoLD} = require('crypto-ld');
+const {Ed25519Signature2020} =
+  require('@digitalbazaar/ed25519-signature-2020');
+const {Ed25519VerificationKey2020} =
+  require('@digitalbazaar/ed25519-verification-key-2020');
 
 const {config, util: {BedrockError}} = bedrock;
 
