@@ -1,11 +1,16 @@
-/*
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
+/*!
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import {config} from '@bedrock/core';
+import {fileURLToPath} from 'url';
+import path from 'path';
+import '@bedrock/https-agent';
+import '@bedrock/kms-http';
+import '@bedrock/mongodb';
 
-const {config} = require('bedrock');
-const path = require('path');
-require('bedrock-express');
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+
+config.mocha.tests.push(path.join(__dirname, 'mocha'));
 
 // MongoDB
 config.mongodb.name = 'bedrock_kms_http_test';
@@ -14,8 +19,6 @@ config.mongodb.dropCollections.collections = [];
 
 // do not require an authentication session for tests
 config['kms-http'].requireAuthentication = false;
-
-config.mocha.tests.push(path.join(__dirname, 'mocha'));
 
 // allow self-signed certs in test framework
 config['https-agent'].rejectUnauthorized = false;
